+++ shuup/admin/modules/products/views/edit.py
@@ -22,6 +22,7 @@ from shuup.admin.modules.products.forms import (
     ProductAttributesForm, ProductBaseForm, ProductImageMediaFormSet,
     ProductMediaFormSet, ShopProductForm
 )
+from shuup.admin.utils.tour import is_tour_complete
 from shuup.admin.utils.views import CreateOrUpdateView
 from shuup.apps.provides import get_provide_objects
 from shuup.core.models import (
@@ -223,6 +224,8 @@ class ProductEditView(SaveFormPartsMixin, FormPartsViewMixin, CreateOrUpdateView
                     orderability_errors.extend(["%s: %s" % (shop.name, _("Product is not available."))])
         context["orderability_errors"] = orderability_errors
         context["product_sections"] = []
+        context["tour_key"] = "product"
+        context["tour_complete"] = is_tour_complete("product")
         product_sections_provides = sorted(get_provide_objects("admin_product_section"), key=lambda x: x.order)
         for admin_product_section in product_sections_provides:
             if admin_product_section.visible_for_object(self.object):